requirements.txt
eventlet
flask
flask-socketio
gunicorn
pymongo
pydrive
requests
requests-html
rq
newspaper3k